requirements/base.txt
Django==2.2
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.12.0
Pillow==8.1.0
psycopg2-binary==2.8.5
pytz==2020.5
sqlparse==0.4.1
